coin_price/requirements.txt
alpaca-py==0.39.0
annotated-types==0.7.0
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
fonttools==4.56.0
Historic-Crypto==0.1.6
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.1
msgpack==1.1.0
numpy==2.2.3
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
PyQt6==6.8.1
PyQt6-Qt6==6.8.2
PyQt6_sip==13.10.0
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
six==1.17.0
sseclient-py==1.8.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
websockets==15.0.1
